+++ nata-server/app.js
@@ -5,10 +5,18 @@ var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 
-var routes = require('./routes/index');
-var users = require('./routes/users');
+
+var socket_io = require('socket.io');
 
 var app = express();
+
+// Socket.io
+var io           = socket_io();
+app.io           = io;
+
+
+var routes = require('./routes/index')(io);
+var users = require('./routes/users');
 
 
 // view engine setup