+++ nata-server/config/env/development.js
@@ -1,0 +1,4 @@
+module.exports = {
+    sessionSecret : 'developmentSessionSecret',
+    db: 'mongodb://localhost/nata'
+};