+++ nata-server/models/Action.js
@@ -1,0 +1,9 @@
+var mongoose = require('mongoose');
+
+
+var ActionSchema = new mongoose.Schema({
+    type: String,
+    description: String
+});
+
+module.exports = mongoose.model('Action', ActionSchema);